# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fanfinity Analytics API
  description: |
    Real-time fan engagement analytics service for processing live match events.

    ## Overview
    This API enables:
    - Ingestion of match events (goals, cards, passes, etc.)
    - Real-time engagement metrics retrieval
    - Health and readiness monitoring
    - Prometheus metrics exposure

    ## Architecture
    Events are ingested via HTTP, durably stored in Kafka, batch-processed by consumers,
    and stored in ClickHouse for real-time analytics queries.

    ## Rate Limits
    The service is designed to handle 1000+ requests/second with sub-200ms latency.
  version: 1.0.0
  contact:
    name: Fanfinity Engineering
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/events:
    post:
      tags:
        - Events
      summary: Ingest a match event
      description: |
        Accepts a match event for asynchronous processing. The event is validated,
        produced to Kafka, and returns immediately with 202 Accepted.

        Events are processed in batches and stored in ClickHouse for analytics.
      operationId: ingestEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
            examples:
              goal:
                summary: Goal event
                value:
                  eventId: "550e8400-e29b-41d4-a716-446655440000"
                  matchId: "match-2024-01-15-001"
                  eventType: "goal"
                  timestamp: "2024-01-15T14:45:00Z"
                  teamId: 1
                  playerId: "player-10"
                  metadata:
                    minute: 45
                    scorer: "Player Name"
              pass:
                summary: Pass event
                value:
                  eventId: "550e8400-e29b-41d4-a716-446655440001"
                  matchId: "match-2024-01-15-001"
                  eventType: "pass"
                  timestamp: "2024-01-15T14:30:00Z"
                  teamId: 2
                  playerId: "player-8"
      responses:
        '202':
          description: Event accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  summary: Invalid JSON
                  value:
                    error: "Bad Request"
                    message: "invalid JSON body"
                invalidUuid:
                  summary: Invalid UUID
                  value:
                    error: "Bad Request"
                    message: "must be a valid UUID"
                    field: "eventId"
        '503':
          description: Service unavailable (Kafka connection issue)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    EventRequest:
      type: object
      required:
        - eventId
        - matchId
        - eventType
        - timestamp
        - teamId
      properties:
        eventId:
          type: string
          format: uuid
          description: Unique identifier for the event
          example: "550e8400-e29b-41d4-a716-446655440000"
        matchId:
          type: string
          description: Unique identifier for the match
          example: "match-2024-01-15-001"
        eventType:
          type: string
          enum:
            - pass
            - shot
            - goal
            - foul
            - yellow_card
            - red_card
            - substitution
            - offside
            - corner
            - free_kick
            - interception
          description: Type of match event
          example: "goal"
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (RFC3339 format)
          example: "2024-01-15T14:45:00Z"
        teamId:
          type: integer
          enum: [1, 2]
          description: Team identifier (1 or 2)
          example: 1
        playerId:
          type: string
          description: Optional player identifier
          example: "player-10"
        metadata:
          type: object
          additionalProperties: true
          description: Optional additional event data
          example:
            minute: 45
            scorer: "Player Name"
    EventResponse:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
          description: The event ID that was accepted
        status:
          type: string
          enum: [accepted]
          description: Status of the event
        timestamp:
          type: string
          format: date-time
          description: When the event was accepted
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Human-readable error message
          example: "must be a valid UUID"
        field:
          type: string
          description: Field that caused the error (if applicable)
          example: "eventId"